pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeatlas"
version = "0.1.0"
description = "A production-ready CLI + TUI tool for scanning, analyzing, cleaning, and interactively editing comments in codebases"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "sarat1kyan", url = "https://github.com/sarat1kyan"}
]
keywords = ["code-analysis", "comments", "codebase", "cli", "tui"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Utilities",
]
dependencies = [
    "rich>=13.7.0",
    "typer>=0.9.0",
    "textual>=0.48.0",
    "pyyaml>=6.0.1",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.7.0",
    "ruff>=0.1.8",
]

[project.scripts]
codeatlas = "codeatlas.cli:app"

[project.urls]
Homepage = "https://github.com/sarat1kyan/CodeAtlas"
Repository = "https://github.com/sarat1kyan/CodeAtlas"
Issues = "https://github.com/sarat1kyan/CodeAtlas/issues"

[tool.setuptools]
packages = ["codeatlas"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=codeatlas --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["codeatlas"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

